raw2dng: fix autodetection of metadata block when reading raw12 from stdin (T1201)
Previous approach: the presence of metadata block was detected from the
file size (which doesn't make sense on stdin, so the check was skipped)

Current approach: always try to read the metadata block. If it fails,
assume no metadata. There should be no extra data in the raw12 file
after the metadata block; error message if any is found.